moved build with closure tests to integration dir
