+++ test/integration/fixtures/build-withClosure/package.json
@@ -1,0 +1,7 @@
+{
+  "scripts": {
+    "build": "tsdx build"
+  },
+  "name": "build-withclosure",
+  "license": "MIT"
+}
