+++ test/integration/fixtures/build-withClosure/src/index.ts
