+++ test/integration/fixtures/build-withClosure/tsdx.config.closure-advanced.js
